rename packages and classes and delete wildcard imports
