+++ src/main/java/com/example/javaHomeworkSecondTerm/Application.java
@@ -1,0 +1,13 @@
+package com.example.javaHomeworkSecondTerm;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+
+}
