+++ src/main/java/com/example/javaHomeworkSecondTerm/exception/DuplicateEmailException.java
@@ -1,0 +1,7 @@
+package com.example.javaHomeworkSecondTerm.exception;
+
+public class DuplicateEmailException extends RuntimeException {
+    public DuplicateEmailException(String email) {
+        super("Пользователь с email %s уже существует".formatted(email));
+    }
+}
