+++ src/main/java/com/example/javaHomeworkSecondTerm/exception/UniversityNotFoundException.java
@@ -1,0 +1,7 @@
+package com.example.javaHomeworkSecondTerm.exception;
+
+public class UniversityNotFoundException extends RuntimeException {
+    public UniversityNotFoundException(Long id) {
+        super("Университет с ID %d не найден".formatted(id));
+    }
+}
