+++ src/main/java/com/example/javaHomeworkSecondTerm/repository/BooksRepository.java
@@ -1,0 +1,6 @@
+package com.example.javaHomeworkSecondTerm.repository;
+
+import com.example.javaHomeworkSecondTerm.model.Book;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BooksRepository extends JpaRepository<Book, Long> { }
