+++ src/main/java/com/example/javaHomeworkSecondTerm/repository/CoursesRepository.java
@@ -1,0 +1,6 @@
+package com.example.javaHomeworkSecondTerm.repository;
+
+import com.example.javaHomeworkSecondTerm.model.Course;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CoursesRepository extends JpaRepository<Course, Long> { }
