+++ src/main/java/com/example/javaHomeworkSecondTerm/repository/UniversitiesRepository.java
@@ -1,0 +1,6 @@
+package com.example.javaHomeworkSecondTerm.repository;
+
+import com.example.javaHomeworkSecondTerm.model.University;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UniversitiesRepository extends JpaRepository<University, Long> { }
