+++ src/main/java/com/example/javaHomeworkSecondTerm/repository/UsersRepository.java
@@ -1,0 +1,6 @@
+package com.example.javaHomeworkSecondTerm.repository;
+
+import com.example.javaHomeworkSecondTerm.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UsersRepository extends JpaRepository<User, Long> { }
