--- src/main/java/com/example/java_homework_second_term/Application.java
@@ -1,13 +1,0 @@
-package com.example.java_homework_second_term;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-
-}
