--- src/main/java/com/example/java_homework_second_term/controller/CourseController.java
@@ -1,34 +1,0 @@
-package com.example.java_homework_second_term.controller;
-
-import com.example.java_homework_second_term.model.Course;
-import com.example.java_homework_second_term.service.CourseService;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.Collection;
-
-@Slf4j
-@RestController
-@RequestMapping("/courses")
-@RequiredArgsConstructor
-public class CourseController {
-    private final CourseService courseService;
-
-    @GetMapping
-    public ResponseEntity<Collection<Course>> getAllCourses() {
-        final Collection<Course> result = courseService.getAllCourses();
-        return ResponseEntity.ok(result);
-    }
-
-    @PostMapping
-    public ResponseEntity<Course> createCourse(@Valid @RequestBody Course course) {
-        final Course result = courseService.createCourse(course);
-        log.info("created %s".formatted(result.toString()));
-        return ResponseEntity.status(HttpStatus.CREATED).body(result);
-    }
-}
-
