--- src/main/java/com/example/java_homework_second_term/exception/UniversityNotFoundException.java
@@ -1,7 +1,0 @@
-package com.example.java_homework_second_term.exception;
-
-public class UniversityNotFoundException extends RuntimeException {
-    public UniversityNotFoundException(Long id) {
-        super("Университет с ID %d не найден".formatted(id));
-    }
-}
