--- src/main/java/com/example/java_homework_second_term/repository/BookRepository.java
@@ -1,6 +1,0 @@
-package com.example.java_homework_second_term.repository;
-
-import com.example.java_homework_second_term.model.Book;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BookRepository extends JpaRepository<Book, Long> { }
