--- src/main/java/com/example/java_homework_second_term/repository/UniversityRepository.java
@@ -1,6 +1,0 @@
-package com.example.java_homework_second_term.repository;
-
-import com.example.java_homework_second_term.model.University;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UniversityRepository extends JpaRepository<University, Long> { }
