+++ src/main/resources/application.yaml
@@ -2,7 +2,7 @@ spring:
   application:
     name: Java homework second term
   datasource:
-    url: jdbc:h2:mem:testdb
+    url: jdbc:h2:mem:testdыДb
     driver-class-name: org.h2.Driver
     username: sa
     password:
