--- src/test/java/com/example/java_homework_second_term/mock/UserControllerTest.java
@@ -1,62 +1,0 @@
-package com.example.java_homework_second_term.mock;
-
-import com.example.java_homework_second_term.controller.UserController;
-import com.example.java_homework_second_term.model.User;
-import com.example.java_homework_second_term.service.UserService;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.servlet.MockMvc;
-
-import java.util.List;
-
-import static org.mockito.Mockito.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@ExtendWith(SpringExtension.class)
-@WebMvcTest(UserController.class)
-public class UserControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @MockBean
-    private UserService userService;
-
-    @Test
-    public void testGetAllUsers() throws Exception {
-        List<User> users = List.of(new User(1L, "test@email.com", "John", "Doe"));
-        when(userService.getAllUsers()).thenReturn(users);
-
-        mockMvc.perform(get("/users"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].email").value("test@email.com"));
-    }
-
-    @Test
-    public void testCreateUser() throws Exception {
-        User user = new User(1L, "test@email.com", "John", "Doe");
-        when(userService.createUser(any())).thenReturn(user);
-
-        mockMvc.perform(post("/users")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"email\": \"test@email.com\", \"name\": \"John\", \"surname\": \"Doe\"}"))
-                .andExpect(status().isCreated())
-                .andExpect(jsonPath("$.email").value("test@email.com"));
-    }
-
-    @Test
-    public void testCreateUser_InvalidInput() throws Exception {
-        mockMvc.perform(post("/users")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{}"))
-                .andExpect(status().isBadRequest());
-    }
-}
