+++ src/test/java/com/example/java_homework_second_term/mock/UsersControllerTest.java
@@ -1,0 +1,63 @@
+package com.example.java_homework_second_term.mock;
+
+import com.example.javaHomeworkSecondTerm.controller.UsersController;
+import com.example.javaHomeworkSecondTerm.model.User;
+import com.example.javaHomeworkSecondTerm.service.UserService;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.test.web.servlet.MockMvc;
+
+import java.util.List;
+
+import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.any;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@ExtendWith(SpringExtension.class)
+@WebMvcTest(UsersController.class)
+public class UsersControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @MockBean
+    private UserService userService;
+
+    @Test
+    public void testGetAllUsers() throws Exception {
+        List<User> users = List.of(new User(1L, "test@email.com", "John", "Doe"));
+        when(userService.getAllUsers()).thenReturn(users);
+
+        mockMvc.perform(get("/users"))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$[0].email").value("test@email.com"));
+    }
+
+    @Test
+    public void testCreateUser() throws Exception {
+        User user = new User(1L, "test@email.com", "John", "Doe");
+        when(userService.createUser(any())).thenReturn(user);
+
+        mockMvc.perform(post("/users")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content("{\"email\": \"test@email.com\", \"name\": \"John\", \"surname\": \"Doe\"}"))
+                .andExpect(status().isCreated())
+                .andExpect(jsonPath("$.email").value("test@email.com"));
+    }
+
+    @Test
+    public void testCreateUser_InvalidInput() throws Exception {
+        mockMvc.perform(post("/users")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content("{}"))
+                .andExpect(status().isBadRequest());
+    }
+}
